% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.simple_logical}
\alias{is.simple_logical}
\title{Checks if a string is a simple string.}
\usage{
is.simple_logical(logical)
}
\arguments{
\item{logical}{ANY. An R object.}
}
\value{
TRUE or FALSE according it is a simple logical.
}
\description{
A logical string is (a) logical, (b) length 1, and (c) not NA.
}

