% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.simple_string}
\alias{is.simple_string}
\title{Checks if a string is a simple string.}
\usage{
is.simple_string(string)
}
\arguments{
\item{string}{ANY. An R object.}
}
\value{
TRUE or FALSE according it is a simple string.
}
\description{
A simple string is (a) character, (b) length 1, and (c) not blank or NA.
}

