% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_validation_plot.R
\name{plot.validation_object_list}
\alias{plot.validation_object_list}
\alias{plot_validation_object_list}
\title{Plot a \code{validation_plot} object.}
\usage{
\method{plot}{validation_object_list}(x, ...)

plot_validation_object_list(validation_object_list, ylab = "Default (\%)",
  title = NULL, linecolors = c(), line_types = NULL, plotroc = TRUE,
  numbers = TRUE, showpoints = FALSE, increment = 10, n_per_row = 1,
  scale = 1, use_normalizer = FALSE)
}
\arguments{
\item{x}{object.}

\item{...}{Internal.}

\item{validation_object_list}{list. The validation objects to plot.}

\item{ylab}{character. (For plot_type = "standard" only. Default is "Default (\%)")}

\item{title}{character. The title of the plot, by default \code{NULL} for
no title.}

\item{linecolors}{character. For plot_type = "standard" only.
Vector of colors for each model. Default is rainbow.}

\item{line_types}{character. "Solid", "dashed" or "dotted", by default the former.}

\item{plotroc}{logical. Whether to print ROC in legend.}

\item{numbers}{logical. Whether to print number of defaults in
each bucket above the plotted point, but only if you are
plotting a single model. Default is \code{TRUE}.}

\item{showpoints}{logical. If \code{TRUE}, print only lines
on validation graph, not points.}

\item{increment}{integer. Increments of y-axis scale, in percent.
Default = 10.}

\item{n_per_row}{integer.}

\item{scale}{integer. If different than 1, used to scale output image.}

\item{use_normalizer}{logical.}
}
\description{
Plot a \code{validation_plot} object.
}

