% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.validationplot.R
\docType{package}
\name{validationplot}
\alias{validationplot}
\alias{validationplot-package}
\title{R plotting tools for visualizing a Hosmer-Lemeshow test on a binary classifier.}
\description{
The performance of any binary classifier that provides a probabilistic
output (rather than a binary value) can be visualized by
splitting the target validation population into quantiles
of a fixed count (usually 10, that is, deciles) and comparing
the empirical incidence of the dependent variable to the
probabilistic response outputted by the classifier.
}
\details{
The validationplot package makes this endeavour straightforward,
especially in conjunction with the \code{tundra} package.
}
\note{
The package was originally authored by physicist Ryland Ely.
}

